Changes based on VC++ compiling.
git-svn-id: svn://svn.code.sf.net/p/pebl/code-0/trunk@166 12333153-ee01-4b9b-8cc9-3b69a7b36d39

+++ src/platforms/sdl/SDLUtility.cpp
@@ -115,7 +115,7 @@ void SDLUtility::DrawLine(SDL_Surface *surface, int x1, int y1, int x2, int y2, 
     
     //We need a rough estimate of the number of steps to take.  
     
-    int length = (int)(sqrt((x1-x2) * (x1 - x2) + (y1-y2)*(y1-y2)) * 2);
+    int length = (int)(sqrt((double)(x1-x2) * (x1 - x2) + (y1-y2)*(y1-y2)) * 2);
     
     //Draw twice for each intermediate pixel.
     float deltax = (float)(x2 - x1) / length;
